pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unstruwwel-py"
version = "0.0.1"
description = "Python port of unstruwwel: detect and parse historic dates to ISO 8601"
readme = "README_PY.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Unstruwwel maintainers" }
]
keywords = ["dates", "iso8601", "historic", "nlp", "parsing"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "lingua-language-detector>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
  "pytest-cov>=4",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-q"
